Address review: backfill timeouts, 500 vs 403 semantics, file split
- migrations/000078: add lock_timeout / statement_timeout to bound the
  backfill against unrelated long-running lock holders.
- middleware/auth: return 500 MEMBERSHIP_RESOLUTION_FAILED for DB errors
  during membership resolution so ops can distinguish infra failures
  from the "not a member" case (which graceful-degrades, not 403).
- handlers/team ChangeRole: soften the post-update display lookup to
  return 200 with a minimal {id, role} payload instead of 500 — the
  role mutation already committed, so a 500 would mislead clients into
  retrying.
- handlers: split signup vs claim helpers into auth_signup.go and
  auth_claim.go so file names match the flow they host.
